EVG-13295 move rest generator into its own package
