+++ cmd/codegen/core/codegen.go
@@ -1,4 +1,4 @@
-package model
+package core
 
 import (
 	"bytes"
