+++ cmd/codegen/core/codegen_test.go
@@ -1,4 +1,4 @@
-package model
+package core
 
 import (
 	"io/ioutil"
