+++ cmd/codegen/core/conversions.go
@@ -1,4 +1,4 @@
-package model
+package core
 
 import (
 	"fmt"
