+++ cmd/codegen/core/templates/array_methods.gotmpl
