+++ cmd/codegen/core/templates/buildfromservice_conversion.gotmpl
