+++ cmd/codegen/core/templates/conversions.gotmpl
