+++ cmd/codegen/core/templates/field.gotmpl
