+++ cmd/codegen/core/templates/file.gotmpl
