+++ cmd/codegen/core/templates/func_call.gotmpl
