+++ cmd/codegen/core/templates/service_methods.gotmpl
