+++ cmd/codegen/core/templates/struct.gotmpl
