+++ cmd/codegen/core/testdata/expected/aliases.go
