+++ cmd/codegen/core/testdata/expected/aliases__converters.go
