+++ cmd/codegen/core/testdata/expected/array_object.go
