+++ cmd/codegen/core/testdata/expected/array_object__converters.go
