+++ cmd/codegen/core/testdata/expected/array_scalar.go
