+++ cmd/codegen/core/testdata/expected/array_scalar__converters.go
