+++ cmd/codegen/core/testdata/expected/basic.go
