+++ cmd/codegen/core/testdata/expected/object_embedded.go
