+++ cmd/codegen/core/testdata/expected/object_embedded__converters.go
