+++ cmd/codegen/core/testdata/expected/object_simple.go
