+++ cmd/codegen/core/testdata/expected/object_simple__converters.go
