+++ cmd/codegen/core/testdata/expected/scalars.go
