+++ cmd/codegen/core/testdata/expected/thirdparty.go
