+++ cmd/codegen/core/testdata/expected/thirdparty__converters.go
