+++ cmd/codegen/core/testdata/expected/type_alias.go
