+++ cmd/codegen/core/testdata/expected/type_alias__converters.go
