+++ cmd/codegen/core/testdata/expected/types.go
