+++ cmd/codegen/core/testdata/schema/aliases.graphql
