+++ cmd/codegen/core/testdata/schema/array_object.graphql
