+++ cmd/codegen/core/testdata/schema/array_scalar.graphql
