+++ cmd/codegen/core/testdata/schema/basic.graphql
