+++ cmd/codegen/core/testdata/schema/config.yml
