+++ cmd/codegen/core/testdata/schema/object_embedded.graphql
