+++ cmd/codegen/core/testdata/schema/object_simple.graphql
