+++ cmd/codegen/core/testdata/schema/scalars.graphql
