+++ cmd/codegen/core/testdata/schema/thirdparty.graphql
