+++ cmd/codegen/core/testdata/schema/type_alias.graphql
