+++ cmd/codegen/core/testdata/schema/types.graphql
